services/unified-service/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
av==15.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
coloredlogs==15.0.1
ctranslate2==4.6.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.1
faster-whisper==1.2.0
filelock==3.19.1
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
numpy==2.3.3
onnxruntime==1.23.0
packaging==25.0
propcache==0.4.1
protobuf==6.32.1
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.3
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
sentry-sdk==2.39.0
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
tokenizers==0.22.1
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.22.0

# Database dependencies
alembic==1.13.3
psycopg2-binary==2.9.10
sqlalchemy==2.0.36

# Authentication dependencies
bcrypt==4.2.1
passlib==1.7.4
python-jose==3.3.0

# Machine Learning dependencies
scikit-learn==1.5.2
